Porting Time Dataspace unit tests from Jamoma 1 - all ofthem pass
The midi TimeUnit has been renamed midinote to avoid disambiguates with the midigain unit of the Gain Dataspace. Both of these units are enlisted in the NativeUnit enumeration in JamomaDataspace.h (currently lines 44-94), and hence need to differ.


IMHO this commit resolves the most important task of #34
That issue also adds “After creating the unit tests, then integrate with the frequency parameters of the filter objects.” This to me sounds like more of an integration test than a unit test, and I believe it is better to make that a separate issue so that #34 can be closed.